REFACTOR :: applicaiton detail
feature 함수들 정리하고 또 정리했읍니다
