+++ src/apis/application/index.ts
@@ -14,7 +14,7 @@ export const getApplicationDetail = async (id: number) => {
   return (await server.get(`/application/${id}`)).data;
 };
 
-export const replyApplication = async (request: { applicationId: number; replyList: REQUEST }) => {
+export const replyApplication = async (request: { applicationId: number; replyList?: REQUEST }) => {
   return (await server.post(`/reply`, request, authorization())).data;
 };
